fix: create folders for routes and controllers, change some response statuses, etc
