+++ package.json
@@ -8,7 +8,7 @@
     "lint": "tslint --project ./tsconfig.json",
     "task-1-1": "nodemon --exec babel-node ./task-1/task1.1.js",
     "task-1-2": "nodemon --exec babel-node ./task-1/task1.2.js",
-    "start": "tsc && nodemon dist/server.js"
+    "start": "tsc && node dist/server.js"
   },
   "repository": {
     "type": "git",
@@ -26,17 +26,19 @@
     "@babel/node": "^7.13.12",
     "@babel/preset-env": "^7.13.12",
     "@types/express": "^4.17.11",
-    "express": "^4.17.1",
+    "@types/node": "^14.14.37",
+    "@types/uuid": "^8.3.0",
     "nodemon": "^2.0.7",
     "tslint": "^6.1.3",
     "typescript": "^4.2.3"
   },
   "dependencies": {
-    "body-parser": "^1.19.0",
     "csvtojson": "^2.0.10",
     "dotenv": "^8.2.0",
+    "express": "^4.17.1",
     "joi": "^17.4.0",
     "password-validator": "^5.1.1",
+    "process": "^0.11.10",
     "uuid": "^8.3.2"
   }
 }
