+++ task-2/helpers.ts
@@ -1,4 +1,6 @@
-const errorResponse = (schemaErrors) => {
+import Joi from "joi";
+
+const errorResponse = (schemaErrors: Joi.ValidationErrorItem[]) => {
     const errors = schemaErrors.map(error => {
         const { path, message } = error;
         return { path, message };
